fix modelo de datos

+++ Dtos/CambioAceiteDto.cs
@@ -5,8 +5,8 @@ namespace ApiSwagger.Dtos
         public int Id { get; set; }
         public int ColectivoId { get; set; }
         public DateOnly Fecha { get; set; }
-    public int Kilometros { get; set; }
-    public bool FiltrosCambiados { get; set; }
-    // ...propiedad Observaciones eliminada...
+        public decimal Kilometros { get; set; } // Cambiado de int a decimal
+        public bool FiltrosCambiados { get; set; }
+        // ...propiedad Observaciones eliminada...
     }
 }
